Rename configuration button component
