+++ src/components/configuration-button/configuration-button.scss
