+++ src/components/dashboard-header/dashboard-header.js
@@ -1,5 +1,5 @@
 import Login from '../login/login.vue'
-import Configuration from '../configuration/configuration.vue'
+import Configuration from '@/components/configuration-button/configuration-button.vue'
 import CustomButton from '../custom-button/custom-button.vue'
 import IconComponent from '@/components/icon/icon-component.vue'
